Rename the Eclipse adapter project, namespaces and IDs
