+++ bundles/tools.vitruv.adapters.eclipse/.classpath
