+++ bundles/tools.vitruv.adapters.eclipse/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>tools.vitruv.framework.domains.ui</name>
+	<name>tools.vitruv.adapters.eclipse</name>
 	<comment></comment>
 	<projects>
 	</projects>
