+++ bundles/tools.vitruv.adapters.eclipse/.settings/org.eclipse.core.resources.prefs
