+++ bundles/tools.vitruv.adapters.eclipse/META-INF/MANIFEST.MF
@@ -1,12 +1,12 @@
 Manifest-Version: 1.0
 Bundle-ManifestVersion: 2
-Bundle-Name: Vitruv Framework UI Monitored Editors
-Bundle-SymbolicName: tools.vitruv.framework.domains.ui;singleton:=true
-Automatic-Module-Name: tools.vitruv.framework.domains.ui
+Bundle-Name: Vitruv Eclipse Tool Adapter
+Bundle-SymbolicName: tools.vitruv.adapters.eclipse;singleton:=true
+Automatic-Module-Name: tools.vitruv.adapters.eclipse
 Bundle-Version: 2.1.0.qualifier
 Bundle-RequiredExecutionEnvironment: JavaSE-11
 Bundle-ActivationPolicy: lazy
-Bundle-Activator: tools.vitruv.framework.domains.ui.Activator
+Bundle-Activator: tools.vitruv.adapters.eclipse.Activator
 Require-Bundle: org.apache.log4j,
  org.eclipse.core.runtime,
  org.eclipse.core.resources,
@@ -15,6 +15,6 @@ Require-Bundle: org.apache.log4j,
  edu.kit.ipd.sdq.commons.util.eclipse,
  edu.kit.ipd.sdq.activextendannotations,
  tools.vitruv.framework.vsum
-Export-Package: tools.vitruv.framework.domains.ui.monitorededitor,
- tools.vitruv.framework.domains.ui.builder
+Export-Package: tools.vitruv.adapters.eclipse.builder,
+ tools.vitruv.adapters.eclipse.monitorededitor
 Bundle-Vendor: vitruv.tools
