+++ bundles/tools.vitruv.adapters.eclipse/build.properties
