+++ bundles/tools.vitruv.adapters.eclipse/plugin.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<?eclipse version="3.4"?>
+<plugin>
+   <extension-point id="builder" name="Vitruv Eclipse Project Builder" schema="schema/builder.exsd"/>
+
+</plugin>
