+++ bundles/tools.vitruv.adapters.eclipse/src/tools/vitruv/adapters/eclipse/monitorededitor/MonitoredEditor.xtend
@@ -1,4 +1,4 @@
-package tools.vitruv.framework.domains.ui.monitorededitor
+package tools.vitruv.adapters.eclipse.monitorededitor
 
 interface MonitoredEditor {
 	/**
