+++ bundles/tools.vitruv.adapters.emf/META-INF/MANIFEST.MF
@@ -14,7 +14,7 @@ Require-Bundle: org.eclipse.core.resources,
  org.apache.log4j,
  org.eclipse.emf.ecore,
  org.eclipse.emf.transaction,
- tools.vitruv.framework.domains.ui;visibility:=reexport,
+ tools.vitruv.adapters.eclipse;visibility:=reexport,
  tools.vitruv.framework.vsum,
  tools.vitruv.change.propagation,
  tools.vitruv.change.composite,
