+++ bundles/tools.vitruv.adapters.emf/plugin.xml
@@ -12,7 +12,7 @@
    </extension>
    <extension
          name="Vitruv EMF Project Builder"
-         point="tools.vitruv.framework.domains.ui.builder">
+         point="tools.vitruv.adapters.eclipse.builder">
       <assignment
       		name="Vitruv EMF Project Builder" 
             builderId="tools.vitruv.adapters.emf.builder.id">
