--- bundles/tools.vitruv.framework.domains.ui/plugin.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<?eclipse version="3.4"?>
-<plugin>
-   <extension-point id="builder" name="Vitruvius Project Builder" schema="schema/builder.exsd"/>
-
-</plugin>
